refactor: 메뉴바 로직 이동
