+++ app/src/main/java/woowacourse/shopping/view/shoppingcart/ShoppingCartActivity.kt
@@ -3,7 +3,9 @@ package woowacourse.shopping.view.shoppingcart
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
+import android.view.MenuItem
 import androidx.activity.viewModels
+import androidx.appcompat.widget.Toolbar
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityShoppingCartBinding
 import woowacourse.shopping.domain.Product
@@ -17,7 +19,7 @@ class ShoppingCartActivity :
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setSubActivityMenuBar(getString(R.string.toolbar_title_cart), binding.toolbar)
+        setMenubar(binding.toolbar as Toolbar)
         viewModel.apply {
             requestProductsPage(0)
             productsLiveData.observe(this@ShoppingCartActivity) { page -> updateRecyclerView(page) }
@@ -26,6 +28,21 @@ class ShoppingCartActivity :
             rvShoppingCartList.adapter = ShoppingCartAdapter(this@ShoppingCartActivity)
             viewModel = this@ShoppingCartActivity.viewModel
             handler = this@ShoppingCartActivity
+        }
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        if (item.itemId == android.R.id.home) {
+            finish()
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
+    private fun setMenubar(toolbar: Toolbar) {
+        setSupportActionBar(toolbar)
+        supportActionBar?.apply {
+            setDisplayHomeAsUpEnabled(true)
+            this.title = MENU_BAR_TAG
         }
     }
 
@@ -51,8 +68,10 @@ class ShoppingCartActivity :
     }
 
     companion object {
+        private const val MENU_BAR_TAG = "Cart"
+
         fun newIntent(context: Context): Intent {
-            return Intent(context, ShoppingCartActivityTemplate::class.java)
+            return Intent(context, ShoppingCartActivity::class.java)
         }
     }
 }
